Temporary custom driver name
Register manila-csi in K8s as "manila.csi.openstack.org" instead of "cephfs.manila.csi.openstack.org". v1 will bring multi-protocol support: there would be only a single manila-csi instance accommodating all share protocols, as opposed to one instance per share protocol as it is now. "manila.csi.openstack.org" will be the new driver name. It's impossible to change the name of a driver that's already deployed in the cluster, so we'd be stuck in the single-protocol mode for good. Keep this patch downstream until they release it.